# Speed up the semantic alignment loop in Javascript